Rename "constructors" to "translators"
