+++ stacker/config/translators.py
